requirements.txt
asgiref==3.8.1
certifi==2024.7.4
charset-normalizer==3.3.2
Deprecated==1.2.14
Django==4.2.15
django-prometheus==2.3.1
djangorestframework==3.15.2
drf-yasg==1.21.7
Faker==25.9.1
googleapis-common-protos==1.63.1
grpcio==1.64.1
gunicorn==22.0.0
idna==3.7
importlib_metadata==7.1.0
inflection==0.5.1
opentelemetry-api==1.25.0
opentelemetry-exporter-jaeger==1.10.0
opentelemetry-exporter-jaeger-proto-grpc==1.10.0
opentelemetry-exporter-jaeger-thrift==1.10.0
opentelemetry-exporter-otlp-proto-common==1.25.0
opentelemetry-exporter-otlp-proto-grpc==1.25.0
opentelemetry-exporter-otlp-proto-http==1.25.0
opentelemetry-instrumentation==0.46b0
opentelemetry-instrumentation-dbapi==0.46b0
opentelemetry-instrumentation-django==0.46b0
opentelemetry-instrumentation-celery==0.46b0
opentelemetry-instrumentation-logging==0.46b0
opentelemetry-instrumentation-psycopg2==0.46b0
opentelemetry-instrumentation-requests==0.46b0
opentelemetry-instrumentation-wsgi==0.46b0
opentelemetry-proto==1.25.0
opentelemetry-sdk==1.25.0
opentelemetry-semantic-conventions==0.46b0
opentelemetry-util-http==0.46b0
packaging==24.1
prometheus_client==0.20.0
protobuf==4.25.3
psycopg2==2.9.9
python-dateutil==2.9.0.post0
pytz==2024.1
PyYAML==6.0.1
requests==2.32.3
six==1.16.0
sqlparse==0.5.0
thrift==0.20.0
typing_extensions==4.12.2
uritemplate==4.1.1
urllib3==2.2.2
uWSGI==2.0.26
wrapt==1.16.0
zipp==3.19.2
django-environ==0.11.2
pytest
celery==5.2.3
django-celery-beat==2.6.0
coverage==7.6.0
selenium==4.23.1